+++ src/components/molecule/form-group/FormGroup.js
@@ -5,10 +5,11 @@ import { Col } from 'react-flexbox-grid';
 
 import css from './FormGroup.module.scss';
 
-const FormGroup = ({ error, success, children, ...otherProps }) => {
+const FormGroup = ({ error, icon, success, children, ...otherProps }) => {
   const classes = classNames({
     [css['c-form-group']]: true,
     [css['c-form-group--error']]: error,
+    [css['c-form-group--error--icon']]: icon,
     [css['c-form-group--success']]: success
   });
 
@@ -21,11 +22,13 @@ const FormGroup = ({ error, success, children, ...otherProps }) => {
 
 FormGroup.propTypes = {
   error: PropTypes.bool,
+  icon: PropTypes.bool,
   success: PropTypes.bool,
   children: PropTypes.oneOfType([PropTypes.object, PropTypes.array])
 };
 FormGroup.defaultProps = {
   error: false,
+  icon: false,
   success: false,
   children: null
 };