+++ src/components/organism/form/Form.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-const Form = ({ onSubmit, children }) => (
-  <form onSubmit={onSubmit} noValidate>
+const Form = ({ onSubmit, className, children }) => (
+  <form onSubmit={onSubmit} className={className} noValidate>
     {children}
   </form>
 );
@@ -13,9 +13,12 @@ Form.propTypes = {
     PropTypes.object,
     PropTypes.string
   ]).isRequired,
-  onSubmit: PropTypes.func.isRequired
+  onSubmit: PropTypes.func.isRequired,
+  className: PropTypes.string
 };
 
-Form.defaultProps = {};
+Form.defaultProps = {
+  className: null
+};
 
 export default Form;