+++ src/index.js
@@ -15,16 +15,18 @@ import ListItem from './components/atom/list-item/ListItem';
 import Pagination from './components/atom/pagination/Pagination';
 import Select from './components/atom/select/Select';
 import Text from './components/atom/text/Text';
+import MaskedInput from './components/atom/masked-input/MaskedInput';
 
 // molecules
 import Card from './components/molecule/card/Card';
+import FilterHeader from './components/molecule/filter-header/FilterHeader';
+import FilterContent from './components/molecule/filter-content/FilterContent';
+import FilterFooter from './components/molecule/filter-footer/FilterFooter';
 import FormActions from './components/molecule/form-actions/FormActions';
 import FormGroup from './components/molecule/form-group/FormGroup';
 import FormRow from './components/molecule/form-row/FormRow';
 import List from './components/molecule/list/List';
-import FilterHeader from './components/molecule/filter-header/FilterHeader';
-import FilterContent from './components/molecule/filter-content/FilterContent';
-import FilterFooter from './components/molecule/filter-footer/FilterFooter';
+import Tooltip from './components/molecule/tooltip/Tooltip';
 
 // organisms
 import Form from './components/organism/form/Form';
@@ -52,8 +54,10 @@ export {
   List,
   ListItem,
   Loader,
+  MaskedInput,
   Pagination,
   Select,
   Text,
+  Tooltip,
   Row
 };